Refactor test yves to latest application version
