+++ tests/App/config/yves.php
@@ -4,7 +4,7 @@ use Spryker\Shared\Application\ApplicationConstants;
 
 $config[ApplicationConstants::YVES_SSL_ENABLED] = false;
 $config[ApplicationConstants::HOST_YVES] = getenv('HOST_YVES') ?: 'localhost';
-$config[ApplicationConstants::PORT_YVES] = getenv('PORT_YVES') ?: '8080';
+$config[ApplicationConstants::PORT_YVES] = getenv('PORT_YVES') ?: '8085';
 $config[ApplicationConstants::BASE_URL_YVES] = sprintf(
     'http://%s:%s',
     $config[ApplicationConstants::HOST_YVES],
